rename and fix modal path
